Fix imports for libnetwork move
We are moving libnetwork to the main docker repo.
Because docker/docker imports docker/swarmkit we need to do this
intermediate step to update swarmkit to use the update imports.

Other vendoring changes are related to changes in docker/docker (like
updated errors package for errors.Is support, and "platforms" package from
containerd, since these are used in the docker API client.)

Signed-off-by: Brian Goff <cpuguy83@gmail.com>